Change from int to double to allow new input method

+++ include/inputModes.hpp
@@ -4,8 +4,8 @@
 #include <vector>
 #include <iostream>
 
-void printWeightsMatrix(const std::vector<std::vector<int> > &weights);
-void matrixModeInput(int n, std::vector<std::vector<int> > &weights, std::istream& stream = std::cin);
-void edgeModeInput(int n, std::vector<std::vector<int> > &weights, std::istream& stream = std::cin);
-void weightsModeInput(int n, std::vector<std::vector<int> > &weights, std::istream& stream = std::cin);
+void printWeightsMatrix(const std::vector<std::vector<double> > &weights);
+void matrixModeInput(int n, std::vector<std::vector<double> > &weights, std::istream& stream = std::cin);
+void edgeModeInput(int n, std::vector<std::vector<double> > &weights, std::istream& stream = std::cin);
+void weightsModeInput(int n, std::vector<std::vector<double> > &weights, std::istream& stream = std::cin);
 void getInputModeManually(std::string& mode);
